feat: add hledger as a valid output format
Right now, I've simply added the format `hledger` to the `OutputFormat` type, which is also used by the `internal.stock.Search` function, despite this function not supporting that output.
Later, it would be nice to separate the output types, but I've yet to find a generic way to implement it.