fix: add condition to truncate decimal TZs
Some search results from the Alpha Vantage API were returning timezones such as `UTC+5.5`, which broke the parsing of the offset. I tried parsing it as a float32, but then the `time.FizedZone` method still would have trouble with the original string staying as `UTC+5.5`, so I simply truncate any decimal timezone and be done with it.